requirements.txt
flask==2.0.1
tensorflow==2.15.0
numpy==1.19.5
pillow==8.3.1
requests==2.26.0
#python==3.8 to 3.10